Rewrite animation export code to better handle complex parent relations.
The old rule "objects with children are placed in a new file" created
too many (empty) files. Therefore, new files are now only created when
it is necessary to properly export an animation.

Each subset in one of the exported files is assigned a subset
identifier, consisting of 1) a custom name for the subser (usually
empty), 2) the name of the material to export, 3) the name of the object
controlling the subset's animation. This ensures that the exported
subsets meet Zusi's limitations: one subset can have only one material
and be controlled by one animation at the same time.

One Blender object may be part of many subsets.

Also re-organize the export tests to make the file more clearly structured.